+++ data/i18n/hydrogen_es.ts
@@ -3013,6 +3013,14 @@ Debería funcionar correctamente mientras utilices el GMRockKit, y no uses tresi
         <source>Switch metronome on/off</source>
         <translation>Encender/apagar metrónomo</translation>
     </message>
+    <message>
+        <source>Count in and play/ Pause</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Counting in</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>MasterLine</name>
@@ -5631,6 +5639,18 @@ p, li { white-space: pre-wrap; }
     </message>
     <message>
         <source>Add component to current instrument</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start playback</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start/pause playback</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start/stop playback</source>
         <translation type="unfinished"></translation>
     </message>
 </context>