+++ data/i18n/hydrogen_pl.ts
@@ -2978,6 +2978,14 @@ It should work like a charm provided that you use the GMRockKit, and that you do
         <source>Switch metronome on/off</source>
         <translation type="unfinished"></translation>
     </message>
+    <message>
+        <source>Count in and play/ Pause</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Counting in</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>MasterLine</name>
@@ -5566,6 +5574,18 @@ p, li { white-space: pre-wrap; }
     </message>
     <message>
         <source>Add component to current instrument</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start playback</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start/pause playback</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start/stop playback</source>
         <translation type="unfinished"></translation>
     </message>
 </context>