+++ data/i18n/hydrogen_zh_CN.ts
@@ -3008,6 +3008,14 @@ LilyPond 导出是一项实验性功能。
         <source>Switch metronome on/off</source>
         <translation>打开/关闭节拍器</translation>
     </message>
+    <message>
+        <source>Count in and play/ Pause</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Counting in</source>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>MasterLine</name>
@@ -5626,6 +5634,18 @@ p, li { white-space: pre-wrap; }
     </message>
     <message>
         <source>Add component to current instrument</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start playback</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start/pause playback</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Count in and start/stop playback</source>
         <translation type="unfinished"></translation>
     </message>
 </context>