+++ src/gui/src/MainToolBar/MainToolBar.cpp
@@ -84,10 +84,10 @@ MainToolBar::MainToolBar( QWidget* pParent) : QToolBar( pParent )
 	};
 
 	auto createButton = [&]( const QString& sText, bool bCheckable ) {
-		auto pAction = new MidiLearnableToolButton( this, sText );
-		pAction->setCheckable( bCheckable );
+		auto pButton = new MidiLearnableToolButton( this, sText );
+		pButton->setCheckable( bCheckable );
 
-		return pAction;
+		return pButton;
 	};
 	////////////////////////////////////////////////////////////////////////////
 
@@ -176,8 +176,46 @@ MainToolBar::MainToolBar( QWidget* pParent) : QToolBar( pParent )
 	connect( m_pPlayButton, &QToolButton::clicked, [&]() {
 		playBtnClicked();
 	} );
+
+	m_pPlayMidiAction = std::make_shared<MidiAction>(
+		MidiAction::Type::PlayPauseToggle );
+	m_pPlayAction = new QAction( this );
+	m_pPlayAction->setCheckable( true );
+	m_pPlayAction->setText( tr( "Play/ Pause" ) );
+	connect( m_pPlayAction, &QAction::triggered, [=]() {
+		auto pPref = Preferences::get_instance();
+		if ( pPref->getCountIn() ) {
+			pPref->setCountIn( false );
+		}
+		if ( m_pPlayButton->defaultAction() != m_pPlayAction ) {
+			m_pPlayButton->setDefaultAction( m_pPlayAction );
+		}
+		m_pPlayButton->setMidiAction( m_pPlayMidiAction );
+		updateTransportControl();
+	} );
+	m_pCountInMidiAction = std::make_shared<MidiAction>(
+		MidiAction::Type::PlayPauseToggle );
+	m_pCountInAction = new QAction( this );
+	m_pCountInAction->setCheckable( true );
+	m_pCountInAction->setText( tr( "Count in and play/ Pause" ) );
+	connect( m_pCountInAction, &QAction::triggered, [=]() {
+		auto pPref = Preferences::get_instance();
+		if ( ! pPref->getCountIn() ) {
+			pPref->setCountIn( true );
+		}
+		if ( m_pPlayButton->defaultAction() != m_pCountInAction ) {
+			m_pPlayButton->setDefaultAction( m_pCountInAction );
+		}
+		m_pPlayButton->setMidiAction( m_pCountInMidiAction );
+		updateTransportControl();
+	} );
+
+	m_pPlayButton->addAction( m_pPlayAction );
+	m_pPlayButton->addAction( m_pCountInAction );
+	m_pPlayButton->setDefaultAction(
+		pPref->getCountIn() ? m_pCountInAction : m_pPlayAction );
 	m_pPlayButton->setMidiAction(
-		std::make_shared<MidiAction>( MidiAction::Type::PlayPauseToggle ) );
+		pPref->getCountIn() ? m_pCountInMidiAction : m_pPlayMidiAction );
 	addWidget( m_pPlayButton );
 
 	// Stop button
@@ -484,7 +522,9 @@ void MainToolBar::metronomeEvent( int nValue ) {
 		return;
 	}
 
-	if ( ! pPref->m_bUseMetronome ) {
+	if ( ! pPref->m_bUseMetronome &&
+		 Hydrogen::get_instance()->getAudioEngine()->getState() !=
+		 AudioEngine::State::CountIn ) {
 		return;
 	}
 
@@ -584,8 +624,10 @@ void MainToolBar::recBtnClicked() {
 
 /// Start audio engine
 void MainToolBar::playBtnClicked() {
-	auto pCommonStrings = HydrogenApp::get_instance()->getCommonStrings();
+	auto pHydrogenApp = HydrogenApp::get_instance();
+	const auto pCommonStrings = pHydrogenApp->getCommonStrings();
 	auto pHydrogen = Hydrogen::get_instance();
+	const auto pPref = Preferences::get_instance();
 
 	// Hint that something is wrong in case there is no proper audio
 	// driver set.
@@ -597,14 +639,20 @@ void MainToolBar::playBtnClicked() {
 							  .arg( pCommonStrings->getAudioDriverErrorHint() ) );
 		return;
 	}
-	
-	if ( m_pPlayButton->isChecked() ) {
-		pHydrogen->sequencerPlay();
-		(HydrogenApp::get_instance())->showStatusBarMessage( tr("Playing.") );
+
+	if ( pHydrogen->getAudioEngine()->getState() != AudioEngine::State::Playing ) {
+		if ( pPref->getCountIn() ) {
+			CoreActionController::startCountIn();
+			pHydrogenApp->showStatusBarMessage( tr( "Counting in" ) );
+		}
+		else {
+			pHydrogen->sequencerPlay();
+			pHydrogenApp->showStatusBarMessage( tr( "Playing." ) );
+		}
 	}
 	else {
 		pHydrogen->sequencerStop();
-		(HydrogenApp::get_instance())->showStatusBarMessage( tr("Pause.") );
+		pHydrogenApp->showStatusBarMessage( tr( "Pause." ) );
 	}
 }
 
@@ -911,7 +959,8 @@ void MainToolBar::updateIcons() {
 	m_pRwdButton->setIcon( QIcon( sIconPath + "rewind.svg" ) );
 	m_pRecButton->setIcon(
 		QIcon( sIconPath + "record.svg" ) );
-	m_pPlayButton->setIcon( QIcon( sIconPath + "play.svg" ) );
+	m_pPlayAction->setIcon( QIcon( sIconPath + "play.svg" ) );
+	m_pCountInAction->setIcon( QIcon( sIconPath + "play_count_in.svg" ) );
 	m_pStopButton->setIcon( QIcon( sIconPath + "stop.svg" ) );
 	m_pFfwdButton->setIcon( QIcon( sIconPath + "fast_forward.svg" ) );
 	m_pSongLoopAction->setIcon( QIcon( sIconPath + "loop.svg" ) );